refactor: extract shared colighter logic
