--- apps/extension/src/constants.ts
@@ -1,11 +1,0 @@
-import { ColorDescription } from './types';
-
-export const HIGHLIGHT_COLOR_OPTIONS: ColorDescription[] = [
-  { name: 'red', val: 'FAA99D' },
-  { name: 'yellow', val: 'FDDF7E' },
-  { name: 'green', val: 'CCE29C' },
-  { name: 'blue', val: '67EBFA' },
-  { name: 'purple', val: 'CE97FB' },
-];
-
-export const DEFAULT_HIGHLIGHT_COLOR = HIGHLIGHT_COLOR_OPTIONS[0];
