+++ packages/colighter/README.md
@@ -1,0 +1,3 @@
+# `colighter`
+
+A collab first highlight experience powered by nostr
