+++ packages/colighter/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "colighter",
+  "version": "0.0.1",
+  "private": false,
+  "license": "MIT",
+  "sideEffects": false,
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "build": "tsup src/index.ts --format cjs --dts",
+    "clean": "rm -rf dist",
+    "dev": "tsup src/index.ts --format cjs --watch --dts",
+    "lint": "eslint \"src/**/*.ts*\""
+  },
+  "dependencies": {
+    "nostrcollab": "*",
+    "nostrfn": "*"
+  },
+  "devDependencies": {
+    "@types/jsrsasign": "^10.5.8",
+    "eslint": "^8.4.1",
+    "eslint-config-custom": "*",
+    "tsconfig": "*",
+    "tsup": "^6.0.0"
+  }
+}
