+++ packages/colighter/src/index.ts
@@ -2,6 +2,9 @@ export {
   type IHighlight,
   type IHighlightCollection,
   type IHighlightCollectionAppModel,
+  type ISerializedRange,
+  HighlightCollectionUpdate,
 } from './types';
 export { HighlightContainerRuntimeFactory } from './container';
 export { Highlight, HighlightCollection } from './model';
+export * from './utils';
