+++ packages/colighter/src/model.ts
@@ -1,16 +1,16 @@
-import { DataObject, DataObjectFactory } from '@fluidframework/aqueduct';
-import { IHighlight, IHighlightCollection, SerializedRange } from './types';
+import { DataObject, DataObjectFactory } from 'nostrcollab';
+import { IHighlight, IHighlightCollection, ISerializedRange } from './types';
 import { sha256Hash } from './utils';
 
 export class Highlight implements IHighlight {
   protected constructor(
     public author: string,
     public text: string,
-    public range: SerializedRange,
+    public range: ISerializedRange,
     public hashId: string
   ) {}
 
-  static async create(text: string, range: SerializedRange, author: string) {
+  static async create(text: string, range: ISerializedRange, author: string) {
     const hash = await sha256Hash(`${author}-${text}`);
     return new Highlight(author, text, range, hash);
   }
