+++ packages/colighter/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "tsconfig/base.json",
+  "include": ["src"],
+  "exclude": ["dist"],
+  "compilerOptions": {
+    "module": "commonjs",
+  }
+}
