+++ packages/nostrcollab/src/index.ts
@@ -1,5 +1,5 @@
 export * from './container';
-export { NostrCollabLoader, StaticCodeLoader } from './loader';
+export { NostrCollabLoader, StaticCodeLoader, loadCollabModel } from './loader';
 export {
   NostrRelayUrlResolver,
   NostrRelayTokenProvider,
@@ -10,3 +10,4 @@ export type { CollabRelay } from './types';
 
 // Type aliases
 export { RouterliciousDocumentServiceFactory } from '@fluidframework/routerlicious-driver';
+export { DataObject, DataObjectFactory } from '@fluidframework/aqueduct';
