+++ packages/serverless/src/functions/Test.ts
@@ -1,15 +1,23 @@
-import { app, HttpRequest, HttpResponseInit, InvocationContext } from "@azure/functions";
+import {
+  app,
+  HttpRequest,
+  HttpResponseInit,
+  InvocationContext,
+} from '@azure/functions';
 
-export async function Test(request: HttpRequest, context: InvocationContext): Promise<HttpResponseInit> {
-    context.log(`Http function processed request for url "${request.url}"`);
+export async function Test(
+  request: HttpRequest,
+  context: InvocationContext
+): Promise<HttpResponseInit> {
+  context.log(`Http function processed request for url "${request.url}"`);
 
-    const name = request.query.get('name') || await request.text() || 'world';
+  const name = request.query.get('name') || (await request.text()) || 'world';
 
-    return { body: `Hello, ${name}!` };
-};
+  return { body: `Hello, ${name}!` };
+}
 
 app.http('Test', {
-    methods: ['GET', 'POST'],
-    authLevel: 'anonymous',
-    handler: Test
+  methods: ['GET', 'POST'],
+  authLevel: 'anonymous',
+  handler: Test,
 });
